title：完成全链路追踪功能 info：     接入Flink实时处理日志信息，将处理结果存入Redis供前端实时查询     处理结果分为用户维度和消息模板维度     用户维度数据【key：userId，value：埋点信息JSON字符串】使用List数据类型存入Redis     消息模板维度数据【key：businessId，field：点位，value：人数】使用Hash数据类型存入Redis

+++ opush-process/src/main/java/com/lin/opush/utils/LoadBalanceUtils.java
@@ -39,7 +39,7 @@ public class LoadBalanceUtils {
     /**
      * 系统动态流量配置【目前暂时支持短信和硬件动态流量配置，此时sendAccount = 0】
      */
-    private static final Integer AUTO_FLOW_RULE = 0;
+    private static final Long AUTO_FLOW_RULE = 0L;
 
     /**
      * 短信不同消息类型对应动态流量配置key
